final commit
Our code invites the user to enter a number into the serial monitor representing each section of the rotating plate in coordination with each angle the motor rotates. We used a stepper.h library to first include the stepper motor commands into the Arduino which enables them to function. The code consists of an algorithm that calculates the angle that the motors rotate, based on the number that is entered and the difference between the current position and the desired one.